Force dtype

+++ pandas/tests/arrays/test_datetimelike.py
@@ -819,6 +819,6 @@ def test_searchsorted_datetimelike_with_listlike(klass):
     # https://github.com/pandas-dev/pandas/issues/32762
     dates = pd.to_datetime(["2020-01-01", "2020-02-01"])
     result = dates.searchsorted(klass(dates))
-    expected = np.array([0, 1])
+    expected = np.array([0, 1], dtype=result.dtype)
 
     tm.assert_numpy_array_equal(result, expected)
